+++ internal/dto/admin/video.go
@@ -6,14 +6,14 @@ import "github.com/ilaziness/orange-tv/internal/dto"
 type VideoListRequest struct {
 	dto.PaginationRequest
 	Keyword       string `form:"keyword"`
-	CategoryID    uint64 `form:"category_id"`
+	CategoryID    uint32 `form:"category_id"`
 	PublishStatus *uint8 `form:"publish_status"`
 	Year          uint32 `form:"year"`
 	Region        string `form:"region"`
 	Language      string `form:"language"`
-	DirectorID    uint64 `form:"director_id"`
-	ActorID       uint64 `form:"actor_id"`
-	TagID         uint64 `form:"tag_id"`
+	DirectorID    uint32 `form:"director_id"`
+	ActorID       uint32 `form:"actor_id"`
+	TagID         uint32 `form:"tag_id"`
 }
 
 // CreateVideoRequest creates a video with associations.
@@ -21,7 +21,7 @@ type CreateVideoRequest struct {
 	Title         string            `json:"title" validate:"required,min=1,max=255"`
 	Subtitle      string            `json:"subtitle" validate:"omitempty,max=255"`
 	Description   string            `json:"description" validate:"omitempty,max=10000"`
-	CategoryID    uint64            `json:"category_id" validate:"required,min=1"`
+	CategoryID    uint32            `json:"category_id" validate:"required,min=1"`
 	PublishStatus *uint8            `json:"publish_status" validate:"omitempty,oneof=0 1"`
 	SerialStatus  *uint8            `json:"serial_status" validate:"omitempty,oneof=1 2 3"`
 	CoverImage    string            `json:"cover_image" validate:"omitempty,max=500"`
@@ -31,9 +31,9 @@ type CreateVideoRequest struct {
 	Duration      uint32            `json:"duration" validate:"omitempty,min=0"`
 	Language      string            `json:"language" validate:"omitempty,max=50"`
 	ReleaseDate   string            `json:"release_date" validate:"omitempty,max=64"`
-	DirectorIDs   []uint64          `json:"director_ids" validate:"omitempty,dive,min=1"`
+	DirectorIDs   []uint32          `json:"director_ids" validate:"omitempty,dive,min=1"`
 	Actors        []VideoActorInput `json:"actors" validate:"omitempty,dive"`
-	TagIDs        []uint64          `json:"tag_ids" validate:"omitempty,dive,min=1"`
+	TagIDs        []uint32          `json:"tag_ids" validate:"omitempty,dive,min=1"`
 }
 
 // UpdateVideoRequest updates a video and optional associations.
@@ -41,7 +41,7 @@ type UpdateVideoRequest struct {
 	Title         *string            `json:"title" validate:"omitempty,min=1,max=255"`
 	Subtitle      *string            `json:"subtitle" validate:"omitempty,max=255"`
 	Description   *string            `json:"description" validate:"omitempty,max=10000"`
-	CategoryID    *uint64            `json:"category_id" validate:"omitempty,min=1"`
+	CategoryID    *uint32            `json:"category_id" validate:"omitempty,min=1"`
 	PublishStatus *uint8             `json:"publish_status" validate:"omitempty,oneof=0 1"`
 	SerialStatus  *uint8             `json:"serial_status" validate:"omitempty,oneof=1 2 3"`
 	CoverImage    *string            `json:"cover_image" validate:"omitempty,max=500"`
@@ -51,19 +51,19 @@ type UpdateVideoRequest struct {
 	Duration      *uint32            `json:"duration" validate:"omitempty,min=0"`
 	Language      *string            `json:"language" validate:"omitempty,max=50"`
 	ReleaseDate   *string            `json:"release_date" validate:"omitempty,max=64"`
-	DirectorIDs   *[]uint64          `json:"director_ids" validate:"omitempty,dive,min=1"`
+	DirectorIDs   *[]uint32          `json:"director_ids" validate:"omitempty,dive,min=1"`
 	Actors        *[]VideoActorInput `json:"actors" validate:"omitempty,dive"`
-	TagIDs        *[]uint64          `json:"tag_ids" validate:"omitempty,dive,min=1"`
+	TagIDs        *[]uint32          `json:"tag_ids" validate:"omitempty,dive,min=1"`
 }
 
 // VideoActorInput binds actor to video.
 type VideoActorInput struct {
-	ActorID uint64 `json:"actor_id" validate:"required,min=1"`
+	ActorID uint32 `json:"actor_id" validate:"required,min=1"`
 }
 
 // VideoListItem is a compact video card payload for admin (includes publish_status, timestamps).
 type VideoListItem struct {
-	ID            uint64          `json:"id"`
+	ID            uint32          `json:"id"`
 	Title         string          `json:"title"`
 	Subtitle      string          `json:"subtitle"`
 	Cover         string          `json:"cover"`
@@ -72,7 +72,7 @@ type VideoListItem struct {
 	Region        string          `json:"region"`
 	Language      string          `json:"language"`
 	Rating        float64         `json:"rating"`
-	CategoryID    uint64          `json:"category_id"`
+	CategoryID    uint32          `json:"category_id"`
 	CategoryName  string          `json:"category_name,omitempty"`
 	PublishStatus uint8           `json:"publish_status"`
 	SerialStatus  uint8           `json:"serial_status"`
@@ -85,7 +85,7 @@ type VideoListItem struct {
 
 // VideoSourceEpisode is one playable episode under a source for admin (includes URL).
 type VideoSourceEpisode struct {
-	ID      uint64 `json:"id"`
+	ID      uint32 `json:"id"`
 	Episode uint32 `json:"episode"`
 	Title   string `json:"title"`
 	URL     string `json:"url"`
@@ -96,18 +96,18 @@ type VideoSourceEpisode struct {
 
 // VideoSourceGroup groups episodes by play source for admin (includes URL).
 type VideoSourceGroup struct {
-	ID       uint64               `json:"id"`
+	ID       uint32               `json:"id"`
 	Name     string               `json:"name"`
 	Episodes []VideoSourceEpisode `json:"episodes"`
 }
 
 // VideoDetailResponse is a full video detail payload for admin (includes publish_status & play URLs).
 type VideoDetailResponse struct {
-	ID            uint64             `json:"id"`
+	ID            uint32             `json:"id"`
 	Title         string             `json:"title"`
 	Subtitle      string             `json:"subtitle"`
 	Description   string             `json:"description"`
-	CategoryID    uint64             `json:"category_id"`
+	CategoryID    uint32             `json:"category_id"`
 	PublishStatus uint8              `json:"publish_status"`
 	SerialStatus  uint8              `json:"serial_status"`
 	Cover         string             `json:"cover"`
